[safesnap] move plugin logic to snapshot-plugins repo
