+++ src/components/Plugin/SafeSnap/Form/Transaction.vue
@@ -1,6 +1,6 @@
 <script>
 import { formatUnits } from '@ethersproject/units';
-import { getAbiFirstFunctionName } from '@/helpers/abi/utils';
+import { getAbiFirstFunctionName } from '@/../snapshot-plugins/src/plugins/safeSnap';
 import { shorten } from '@/helpers/utils';
 
 const labels = {
